API/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asn1crypto==1.5.1
boto3==1.40.51
botocore==1.40.51
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.2
fastapi==0.119.0
filelock==3.20.0
h11==0.16.0
idna==3.11
jmespath==1.0.1
joblib==1.5.2
kafka-python==2.2.15
minio==7.2.18
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
pyarrow==18.1.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.1
pydantic_core==2.41.3
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==4.0.0
sortedcontainers==2.4.0
starlette==0.48.0
threadpoolctl==3.6.0
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
